Fixed issues preventing unit tests from running; Removed unit tests that weren't needed (some had already been tested in other components)
